refactor: split cmd/indexer/main.go into per-concern files
